[Fix]: files moved to Platform/Windows
